Rename file to root-update
