+++ src/langium-model-server/lms/model/index.ts
@@ -5,6 +5,5 @@ export * from './response'
 export * from './semantic-id-response'
 export * from './update/array-update'
 export * from './update/element-update'
-export * from './update/model-update'
+export * from './update/root-update'
 export * from './update/update'
-
